+++ src/pager.rs
@@ -13,7 +13,7 @@ use crate::future::IsSomeAnd;
 use crate::highlight::{highlight_shell, HighlightRole, HighlightSpec};
 use crate::libc::MB_CUR_MAX;
 use crate::operation_context::OperationContext;
-use crate::screen::{wcswidth_rendered, wcwidth_rendered, Line, ScreenData};
+use crate::screen::{wcswidth_rendered, wcwidth_rendered, CharOffset, Line, ScreenData};
 use crate::termsize::Termsize;
 use crate::wchar::prelude::*;
 use crate::wcstringutil::string_fuzzy_match_string;
@@ -253,9 +253,7 @@ impl Pager {
         assert!(stop_row <= row_count);
         assert!(stop_row - start_row <= term_height);
         // This always printed at the end of the command line.
-        let offset_in_cmdline = usize::MAX;
         self.completion_print(
-            offset_in_cmdline,
             cols,
             &width_by_column,
             start_row,
@@ -300,7 +298,7 @@ impl Pager {
                 HighlightRole::pager_progress,
             );
             print_max(
-                offset_in_cmdline,
+                CharOffset::None,
                 &progress_text,
                 spec,
                 term_width,
@@ -329,7 +327,7 @@ impl Pager {
 
         let mut search_field_remaining = term_width - 1;
         search_field_remaining -= print_max(
-            offset_in_cmdline,
+            CharOffset::None,
             wgettext!(SEARCH_FIELD_PROMPT),
             HighlightSpec::new(),
             search_field_remaining,
@@ -337,7 +335,7 @@ impl Pager {
             search_field,
         );
         search_field_remaining -= print_max(
-            offset_in_cmdline,
+            CharOffset::None,
             &search_field_text,
             underline,
             search_field_remaining,
@@ -407,7 +405,6 @@ impl Pager {
     /// \param lst The list of completions to print
     fn completion_print(
         &self,
-        offset_in_cmdline: usize,
         cols: usize,
         width_by_column: &[usize; PAGER_MAX_COLS],
         row_start: usize,
@@ -437,7 +434,7 @@ impl Pager {
 
                 // Print this completion on its own "line".
                 let mut line = self.completion_print_item(
-                    offset_in_cmdline,
+                    CharOffset::Pager(idx),
                     prefix,
                     el,
                     col_width,
@@ -447,7 +444,7 @@ impl Pager {
 
                 // If there's more to come, append two spaces.
                 if col + 1 < cols {
-                    line.append_str(PAGER_SPACER_STRING, HighlightSpec::new(), offset_in_cmdline);
+                    line.append_str(PAGER_SPACER_STRING, HighlightSpec::new(), CharOffset::None);
                 }
 
                 // Append this to the real line.
@@ -462,7 +459,7 @@ impl Pager {
     /// Print the specified item using at the specified amount of space.
     fn completion_print_item(
         &self,
-        offset_in_cmdline: usize,
+        offset_in_cmdline: CharOffset,
         prefix: &wstr,
         c: &PagerComp,
         width: usize,
@@ -1100,7 +1097,7 @@ fn divide_round_up(numer: usize, denom: usize) -> usize {
 /// \param has_more if this flag is true, this is not the entire string, and the string should be
 /// ellipsized even if the string fits but takes up the whole space.
 fn print_max_impl(
-    offset_in_cmdline: usize,
+    offset_in_cmdline: CharOffset,
     s: &wstr,
     color: impl Fn(usize) -> HighlightSpec,
     max: usize,
@@ -1136,7 +1133,7 @@ fn print_max_impl(
 }
 
 fn print_max(
-    offset_in_cmdline: usize,
+    offset_in_cmdline: CharOffset,
     s: &wstr,
     color: HighlightSpec,
     max: usize,